requirements.txt
anyio==3.7.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
Django==4.2.3
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
phonenumbers==8.13.17
Pillow==10.0.0
pkg_resources==0.0.0
psycopg2-binary==2.9.6
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-telegram-bot==20.4
sniffio==1.3.0
sqlparse==0.4.4
TgCrypto==1.2.5
typing_extensions==4.7.1
